renamed compare files without .cmp extention to make comparison easier with file comparison tools (like BeyondCompare) easier update compare files to make current templates

+++ src/ElementSet/Alpha/Test/Sanity/Sql/proc/spTestSqlEntities_Insert.proc.sql
@@ -7,7 +7,7 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE PROCEDURE spTestSqlEntities_Insert
-	@float	float(10),
+	@float	float(0),
 	@datetime	datetime,
 	@bit	bit,
 	@smallint	smallint,
